+++ yellowstone-grpc-geyser/src/plugin.rs
@@ -175,7 +175,7 @@ impl GeyserPlugin for Plugin {
         &self,
         slot: u64,
         parent: Option<u64>,
-        status: SlotStatus,
+        status: &SlotStatus,
     ) -> PluginResult<()> {
         self.with_inner(|inner| {
             let message = Message::Slot(MessageSlot::from_geyser(slot, parent, status));